I recognized that I required the same sorting logic in the Bookshelf component, so I've abstracted the logic to a utility file to import where needed
